_thisisdrachenwald/requirements.txt
feedparser==6.0.0b3
bs4
requests
pyyaml
Pillow
feedgen
pytz